+++ webapp/views/App/views/Data/Explorer/DataQuery/components/dataQueryView.scss
