+++ webapp/views/App/views/Data/Explorer/DataQuery/components/sort/sortEditor.scss
