+++ webapp/views/App/views/Data/Explorer/DataQuery/components/table.scss
