+++ webapp/views/App/views/Data/Explorer/DataQuery/components/tableContent/tableColumnEdit.js
