+++ webapp/views/App/views/Data/Explorer/DataQuery/components/tableHeader/tableHeader.scss
