+++ webapp/views/App/views/Data/Explorer/DataQuery/index.js
@@ -1,0 +1,1 @@
+export { default } from './DataQuery'
