+++ webapp/views/App/views/Data/Explorer/DataQuery/reducer.js
