+++ webapp/views/App/views/Data/Explorer/DataQuery/state/defaults.js
