+++ webapp/views/App/views/Data/Explorer/DataQuery/state/table.js
