--- webapp/views/App/views/Data/Explorer/dataQuery/index.js
@@ -1,1 +1,0 @@
-export { default } from './dataQueryView'
